Moving tests for buffer schema.
